Fixed branch instructions not using offset as a signed byte

+++ cpu6502/instructions/branch.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 import cpu6502.instructions
 
 
@@ -11,7 +13,7 @@ class BCC(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if not self.cpu.ps['carry_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -31,7 +33,7 @@ class BCS(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if self.cpu.ps['carry_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -51,7 +53,7 @@ class BEQ(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if self.cpu.ps['zero_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -71,7 +73,7 @@ class BMI(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if self.cpu.ps['negative_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -91,7 +93,7 @@ class BNE(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if not self.cpu.ps['zero_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -111,7 +113,7 @@ class BPL(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if not self.cpu.ps['negative_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -131,7 +133,7 @@ class BVC(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if not self.cpu.ps['overflow_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
@@ -151,7 +153,7 @@ class BVS(cpu6502.instructions.AbstractInstruction):
     def relative(self):
         if self.cpu.ps['overflow_flag']:
             offset = int(self.cpu.fetch_byte(), base=0)
-            target_address = self.cpu.pc + offset
+            target_address = self.cpu.pc + np.byte(offset)
             ~self.cpu.clock
             if (target_address >> 8) != (self.cpu.pc >> 8):
                 ~self.cpu.clock
